requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langgraph==0.2.0
chromadb==0.5.0
streamlit==1.38.0
pypdf==4.3.1
openai==1.50.0
python-dotenv==1.0.1
tiktoken==0.7.0
duckduckgo-search==6.2.13
